+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Configuration to exclude custom doktypes when automatically generating the breadcrumb (#84)
+
 ## [2.2.2] - 2022-01-02
 
 ### Fixed